Make node-api `no_std` compatible (#253)
* [test-no-std] test node-api and substrate-api-client separately

* [node-api] adjust toml for no-std support

* [node-api/storage] make no_std compatible

* [extrinsic_params] make `no_std` compatible

* [node-api] no-std PoC

* [node-api/error] cleanup

* [node-api/storage] cleanup unused import

* [node-api/events] cleanup unused imports

* [node-api/metadata] fix imports

* [node-api] fix imports

* [node-api/pretty_format] extract formatting of `PalletMetadata` and only format it, don't print to have no_std compatibility.

* [node-api/metadata] re-enable json pretty format

* Revert "[node-api/pretty_format] extract formatting of `PalletMetadata` and only format it, don't print to have no_std compatibility."

This reverts commit 8962d851f466460c21ae947627a092cee562b8bd.

* [node-api/metadata] extract print functions to `print_metadata.rs`

* [node-api/metadata] remove leftovers from thisError

* [node-api] fix std build

* fixing std build

* [node-api] toml fmt

* [node-api/events] fix obsolete comments

* [test-no-std] toml fmt

* [test-no-std] toml fmt

* [node-api] derive more for structs

* [node-api] derive encode and decode where possible

* [node-api] enable missing std features

* [node-api] switch imports to alloc
