+++ node-api/src/lib.rs
@@ -13,16 +13,26 @@
 
 //! Contains stuff to instantiate communication with a substrate node.
 
-use codec::Decode;
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(not(feature = "std"))]
+extern crate alloc;
+#[cfg(not(feature = "std"))]
+use alloc::{borrow::ToOwned, vec::Vec};
+
+use codec::{Decode, Encode};
 
 pub mod error;
 pub mod events;
 pub mod metadata;
 pub mod storage;
 
+#[cfg(feature = "std")]
+mod print_metadata;
+
 /// Wraps an already encoded byte vector, prevents being encoded as a raw byte vector as part of
 /// the transaction payload
-#[derive(Clone, Debug, Eq, PartialEq)]
+#[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd)]
 pub struct Encoded(pub Vec<u8>);
 
 impl codec::Encode for Encoded {
@@ -32,7 +42,7 @@ impl codec::Encode for Encoded {
 }
 
 /// A phase of a block's execution.
-#[derive(Clone, Debug, Eq, PartialEq, Decode)]
+#[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Encode, Decode)]
 pub enum Phase {
     /// Applying an extrinsic.
     ApplyExtrinsic(u32),
