+++ node-api/src/metadata.rs
@@ -20,62 +20,64 @@
 //! This file is mostly subxt.
 
 use crate::{storage::GetStorage, Encoded};
-use codec::{Encode, Error as CodecError};
+use codec::{Decode, Encode, Error as CodecError};
 use frame_metadata::{
     PalletConstantMetadata, RuntimeMetadata, RuntimeMetadataLastVersion, RuntimeMetadataPrefixed,
     StorageEntryMetadata, META_RESERVED,
 };
-use frame_support::serde::Serialize;
 use scale_info::{form::PortableForm, Type, Variant};
 use sp_core::storage::StorageKey;
-use std::{collections::HashMap, convert::TryFrom};
+
+#[cfg(feature = "std")]
+use serde::Serialize;
+
+// We use `BTreeMap` because we can't use `HashMap` in `no_std`.
+use sp_std::collections::btree_map::BTreeMap;
+
+#[cfg(not(feature = "std"))]
+use alloc::{
+    string::{String, ToString},
+    vec,
+    vec::Vec,
+};
 
 /// Metadata error.
-#[derive(Debug, thiserror::Error)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum MetadataError {
     /// Module is not in metadata.
-    #[error("Pallet {0} not found")]
     PalletNotFound(String),
     /// Pallet is not in metadata.
-    #[error("Pallet index {0} not found")]
     PalletIndexNotFound(u8),
     /// Call is not in metadata.
-    #[error("Call {0} not found")]
     CallNotFound(&'static str),
     /// Event is not in metadata.
-    #[error("Pallet {0}, Event {0} not found")]
     EventNotFound(u8, u8),
-    /// Event is not in metadata.
-    #[error("Pallet {0}, Error {0} not found")]
+    /// Error is not in metadata.
     ErrorNotFound(u8, u8),
     /// Storage is not in metadata.
-    #[error("Storage {0} not found")]
     StorageNotFound(&'static str),
     /// Storage type does not match requested type.
-    #[error("Storage type error")]
     StorageTypeError,
-    #[error("Map value type error")]
+    /// Map value type does not match requested type.
     MapValueTypeError,
-    /// Default error.
-    #[error("Failed to decode default: {0}")]
+    /// Failed to decode the value's default.
     DefaultError(CodecError),
     /// Failure to decode constant value.
-    #[error("Failed to decode constant value: {0}")]
     ConstantValueError(CodecError),
     /// Constant is not in metadata.
-    #[error("Constant {0} not found")]
     ConstantNotFound(&'static str),
-    #[error("Type {0} missing from type registry")]
+    /// Type is missing from type registry.
     TypeNotFound(u32),
 }
 
 /// Runtime metadata.
-#[derive(Clone, Debug)]
+#[derive(Clone)]
+#[cfg_attr(feature = "std", derive(Debug))]
 pub struct Metadata {
-    metadata: RuntimeMetadataLastVersion,
-    pallets: HashMap<String, PalletMetadata>,
-    events: HashMap<(u8, u8), EventMetadata>,
-    errors: HashMap<(u8, u8), ErrorMetadata>,
+    pub metadata: RuntimeMetadataLastVersion,
+    pub pallets: BTreeMap<String, PalletMetadata>,
+    pub events: BTreeMap<(u8, u8), EventMetadata>,
+    pub errors: BTreeMap<(u8, u8), ErrorMetadata>,
 }
 
 impl Metadata {
@@ -142,6 +144,7 @@ impl Metadata {
         &self.metadata
     }
 
+    #[cfg(feature = "std")]
     pub fn pretty_format(metadata: &RuntimeMetadataPrefixed) -> Option<String> {
         let buf = Vec::new();
         let formatter = serde_json::ser::PrettyFormatter::with_indent(b" ");
@@ -149,108 +152,16 @@ impl Metadata {
         metadata.serialize(&mut ser).unwrap();
         String::from_utf8(ser.into_inner()).ok()
     }
-
-    pub fn print_overview(&self) {
-        let mut string = String::new();
-        for (name, pallet) in &self.pallets {
-            string.push_str(name.as_str());
-            string.push('\n');
-            for storage in pallet.storage.keys() {
-                string.push_str(" s  ");
-                string.push_str(storage.as_str());
-                string.push('\n');
-            }
-
-            for call in pallet.calls.keys() {
-                string.push_str(" c  ");
-                string.push_str(call.as_str());
-                string.push('\n');
-            }
-
-            for constant in pallet.constants.keys() {
-                string.push_str(" cst  ");
-                string.push_str(constant.as_str());
-                string.push('\n');
-            }
-
-            for event in self.events(pallet.index) {
-                string.push_str(" e  ");
-                string.push_str(&event.event);
-                string.push('\n');
-            }
-
-            for error in self.errors(pallet.index) {
-                string.push_str(" err  ");
-                string.push_str(&error.error);
-                string.push('\n');
-            }
-        }
-
-        println!("{}", string);
-    }
-
-    pub fn print_pallets(&self) {
-        for m in self.pallets.values() {
-            m.print()
-        }
-    }
-
-    pub fn print_pallets_with_calls(&self) {
-        for m in self.pallets.values() {
-            if !m.calls.is_empty() {
-                m.print_calls();
-            }
-        }
-    }
-    pub fn print_pallets_with_constants(&self) {
-        for m in self.pallets.values() {
-            if !m.constants.is_empty() {
-                m.print_constants();
-            }
-        }
-    }
-    pub fn print_pallet_with_storages(&self) {
-        for m in self.pallets.values() {
-            if !m.storage.is_empty() {
-                m.print_storages();
-            }
-        }
-    }
-
-    pub fn print_pallets_with_events(&self) {
-        for pallet in self.pallets.values() {
-            println!(
-                "----------------- Events for Pallet: {} -----------------\n",
-                pallet.name
-            );
-            for m in self.events(pallet.index) {
-                m.print();
-            }
-            println!();
-        }
-    }
-
-    pub fn print_pallets_with_errors(&self) {
-        for pallet in self.pallets.values() {
-            println!(
-                "----------------- Errors for Pallet: {} -----------------\n",
-                pallet.name
-            );
-            for m in self.errors(pallet.index) {
-                m.print();
-            }
-            println!();
-        }
-    }
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone)]
+#[cfg_attr(feature = "std", derive(Debug))]
 pub struct PalletMetadata {
     pub index: u8,
     pub name: String,
-    pub calls: HashMap<String, u8>,
-    pub storage: HashMap<String, StorageEntryMetadata<PortableForm>>,
-    pub constants: HashMap<String, PalletConstantMetadata<PortableForm>>,
+    pub calls: BTreeMap<String, u8>,
+    pub storage: BTreeMap<String, StorageEntryMetadata<PortableForm>>,
+    pub constants: BTreeMap<String, PalletConstantMetadata<PortableForm>>,
 }
 
 impl PalletMetadata {
@@ -285,61 +196,13 @@ impl PalletMetadata {
             .get(key)
             .ok_or(MetadataError::ConstantNotFound(key))
     }
-
-    pub fn print(&self) {
-        println!(
-            "----------------- Pallet: '{}' -----------------\n",
-            self.name
-        );
-        println!("Pallet id: {}", self.index);
-
-        //self.print_calls();
-    }
-
-    pub fn print_calls(&self) {
-        println!(
-            "----------------- Calls for Pallet: {} -----------------\n",
-            self.name
-        );
-        for (name, index) in &self.calls {
-            println!("Name: {}, index {}", name, index);
-        }
-        println!();
-    }
-
-    pub fn print_constants(&self) {
-        println!(
-            "----------------- Constants for Pallet: {} -----------------\n",
-            self.name
-        );
-        for (name, constant) in &self.constants {
-            println!(
-                "Name: {}, Type {:?}, Value {:?}",
-                name, constant.ty, constant.value
-            );
-        }
-        println!();
-    }
-    pub fn print_storages(&self) {
-        println!(
-            "----------------- Storages for Pallet: {} -----------------\n",
-            self.name
-        );
-        for (name, storage) in &self.storage {
-            println!(
-                "Name: {}, Modifier: {:?}, Type {:?}, Default {:?}",
-                name, storage.modifier, storage.ty, storage.default
-            );
-        }
-        println!();
-    }
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Encode, Decode)]
 pub struct EventMetadata {
-    pallet: String,
-    event: String,
-    variant: Variant<PortableForm>,
+    pub pallet: String,
+    pub event: String,
+    pub variant: Variant<PortableForm>,
 }
 
 impl EventMetadata {
@@ -357,19 +220,13 @@ impl EventMetadata {
     pub fn variant(&self) -> &Variant<PortableForm> {
         &self.variant
     }
-
-    pub fn print(&self) {
-        println!("Name: {}", self.event());
-        println!("Variant: {:?}", self.variant());
-        println!()
-    }
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq, Eq, Ord, PartialOrd, Encode, Decode)]
 pub struct ErrorMetadata {
-    pallet: String,
-    error: String,
-    variant: Variant<PortableForm>,
+    pub pallet: String,
+    pub error: String,
+    pub variant: Variant<PortableForm>,
 }
 
 impl ErrorMetadata {
@@ -387,23 +244,15 @@ impl ErrorMetadata {
     pub fn description(&self) -> &[String] {
         self.variant.docs()
     }
-
-    pub fn print(&self) {
-        println!("Name: {}", self.error());
-        println!("Description: {:?}", self.description());
-        println!()
-    }
 }
 
-#[derive(Debug, thiserror::Error)]
+#[derive(Clone, Debug, PartialEq, Eq, Ord, PartialOrd, Encode, Decode)]
 pub enum InvalidMetadataError {
-    #[error("Invalid prefix")]
     InvalidPrefix,
-    #[error("Invalid version")]
     InvalidVersion,
-    #[error("Type {0} missing from type registry")]
+    /// Type is missing from type registry.
     MissingType(u32),
-    #[error("Type {0} was not a variant/enum type")]
+    /// Type was not variant/enum type.
     TypeDefNotVariant(u32),
 }
 
@@ -434,7 +283,7 @@ impl TryFrom<RuntimeMetadataPrefixed> for Metadata {
             .pallets
             .iter()
             .map(|pallet| {
-                let calls = pallet.calls.as_ref().map_or(Ok(HashMap::new()), |call| {
+                let calls = pallet.calls.as_ref().map_or(Ok(BTreeMap::new()), |call| {
                     let type_def_variant = get_type_def_variant(call.ty.id())?;
                     let calls = type_def_variant
                         .variants()
@@ -444,7 +293,7 @@ impl TryFrom<RuntimeMetadataPrefixed> for Metadata {
                     Ok(calls)
                 })?;
 
-                let storage = pallet.storage.as_ref().map_or(HashMap::new(), |storage| {
+                let storage = pallet.storage.as_ref().map_or(BTreeMap::new(), |storage| {
                     storage
                         .entries
                         .iter()
